Cast to 'double' in pow.
git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/trunk/geant4@18896 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ source/processes/hadronic/models/coherent_elastic/src/G4IntegrHadrNucleus.cc
@@ -42,7 +42,7 @@
 
 //          R0      =  sqrt(0.64);
 
-          Rnucl   = R0*pow(static_cast<float>(Anucleus),0.3333);            //{In Fermi }
+          Rnucl   = R0*pow(static_cast<double>(Anucleus),0.3333);            //{In Fermi }
 
         if(Anucleus == 4) Rnucl = 1.2;
